Remove type 'any' and duplicate filter opts #3 #4
